requirements.txt
# PyScale Dependencies

# Required
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
psutil>=5.9.0

# Optional (GPU Acceleration)
torch>=2.0.0
torchvision>=0.15.0
realesrgan>=0.3.0

# Optional (Screen Capture)
mss>=9.0.0